Custom 7.4v Lithium Ion charger + battery indicator and 1.9v, 3.3v, and 5v buck regulator board.
This is a custom battery charger designed for a Nintendo Gamecube
Portable. The design however is very universal as it is a simple lithium
Ion charger and buck regulator board. There is a low battery indicator
that is 2 transistor analog voltage measurement based for simplicity.
This may or may not be the final version as there are still heat issues
being worked out.